+++ flink-yarn/src/main/java/org/apache/flink/yarn/YarnResourceManager.java
@@ -680,7 +680,7 @@ public class YarnResourceManager extends ResourceManager<YarnWorkerNode> impleme
 
 		taskManagerConfig.setFloat(TaskManagerOptions.NETWORK_BUFFERS_MEMORY_FRACTION,
 				flinkConfig.getFloat(TaskManagerOptions.NETWORK_BUFFERS_MEMORY_FRACTION));
-		long networkBufBytes = tmResource.getNetworkMemorySize() << 20;
+		long networkBufBytes = ((long) tmResource.getNetworkMemorySize()) << 20;
 		taskManagerConfig.setLong(TaskManagerOptions.NETWORK_BUFFERS_MEMORY_MIN, networkBufBytes);
 		taskManagerConfig.setLong(TaskManagerOptions.NETWORK_BUFFERS_MEMORY_MAX, networkBufBytes);
